Development

+++ src/component/Test.js
@@ -1,16 +1,17 @@
 import React, { useState, useRef, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
+import { Flipper, Flipped } from "react-flip-toolkit";
 
 import { cellActions } from "../actions/cellActions";
 import Navbar from "./Navbar";
 import Cell from "./Cell";
+import Status from "./Status";
 
 // You can now get a ref directly to the DOM button:
 
 const Test = () => {
   const [focus, setCurrentFocus] = useState(null);
   // const [cells, setCells] = useState([1064, 68414, 6598])
-  const state = useSelector((state) => state);
   const cells = useSelector((state) => state.cells);
   const cellsRef = useRef();
   cellsRef.current = [];
@@ -24,11 +25,24 @@ const Test = () => {
   };
 
   const addCell = () => {
-    dispatch(cellActions.addCell(focus));
+    const state = {
+      value: "hello everyone",
+      results: [],
+      selectedResultIndex: 0
+    }
+    dispatch(cellActions.addCell(state))
+    console.log('hello')
+  }
+
+  const addEmptyCell = () => {
+    dispatch(cellActions.addEmptyCell(focus));
   };
 
   const deleteAll = () => {
-    dispatch(cellActions.deleteAll());
+    const res = window.confirm("Are you sure you want to delete all ?")
+    if (res) {
+      dispatch(cellActions.deleteAll());
+    }
   };
 
   const runAll = () => {
@@ -55,31 +69,37 @@ const Test = () => {
     }
   };
 
+  
+
   return (
     <div className="container">
       <Navbar />
       <nav>
-        <button onClick={addCell}>New Cell</button>
+        <button onClick={addEmptyCell}>New Cell</button>
         <button onClick={deleteAll}>Delete All</button>
         <button onClick={runAll}>Run All</button>
         <button onClick={maximizeAll}>Maximize All</button>
         <button onClick={minimizeAll}>Minimize All</button>
+        <Status />
       </nav>
       <div className="main-section">
-        
-        {cells.map((item, i) => (
-          <Cell
-            key={item.id}
-            id={item.id}
-            item={item}
-            ref={addToRef}
-            onFocusChange={onFocusChange}
-          />
-        ))}
-        {/* <Cell ref={cellsRef} id="asd" /> */}
-      
+        <Flipper flipKey={cells} spring="gentle">
+          {cells.map((item) => (
+            <Flipped key={item.id} flipId={item.id}>
+              <div>
+                <Cell
+                  key={item.id}
+                  id={item.id}
+                  item={item}
+                  ref={addToRef}
+                  onFocusChange={onFocusChange}
+                />
+              </div>
+            </Flipped>
+          ))}
+        </Flipper>
+
       </div>
-      <div className="status">Ready <i className="fa fa-check"></i></div>
     </div>
   );
 };
